Create calculator object and add button interactivity.
In index.html:
-Changed result label to input element and renamed its class
result.text.

In style.html:
-Updated result-label class name in selector to result-text.

In script.js:
-Created a calculator object with 4 datafields: one for the current
number being typed and displayed, one to hold onto the expressions
to be evaluated, and two references to elements, one for the button
grid and one for the result-text.

Added methods to set event listener on the parent container holding
the buttons.

buttonPress, the event handler passed into the event listener, while
an object method, cannot make direct references to the object. I
chose to keep it an object method even though it must reference
the object using a global reference because it makes sense
conceptually to have it contained with the calculator.

I also added some basic methods that the logic of each type of button.